pcb: Connect dot board pin 14 to COIL_PULSE (pin 12, pin 13) too
The 18V supply on this pin doesn't seem to be used, as it's wired to column darlington arrays while the dot boards seem to use the row darlingtons to source/sink. In any case, it should be safe to connect, tested splicing pin 12+13+14 together on a real dot board and it worked fine.